fix bugs

+++ unterwegs/tasks/uploader.py
@@ -1,4 +1,10 @@
 import os
+import time
+import hashlib
+import redis
+
+import unterwegs.tasks.splitter as spltr
+import unterwegs.tasks.indexer as indxr
 
 from celery import shared_task
 from celery.utils.log import get_task_logger
@@ -7,15 +13,40 @@ from pyseaweed import WeedFS
 logger = get_task_logger(__name__)
 
 wd = WeedFS("master", 9333) # weed-fs master address and port
+rd = redis.Redis(host='redis', port=6379, db=1)
 
 
-@shared_task
+def excerpt_md5hash(fname):
+    with open(fname, "rb") as f:
+        file_hash = hashlib.md5()
+        while chunk := f.read(8192):
+            file_hash.update(chunk)
+    return file_hash.hexdigest()
+
+
+@shared_task(
+    max_retries=3,
+    soft_time_limit=5
+)
 def fire(fname):
-    logger.info("start uploading %s" % fname)
-    fid = wd.upload_file(fname)
-    furl = wd.get_file_url(fid)
-    logger.info("finish uploading %s: %s, %s" % (fname, fid, furl))
+    hash = excerpt_md5hash(fname)
+    if rd.hexists('hash', hash) == 0:
+        logger.info("uploading %s" % fname)
+        fid = wd.upload_file(fname)
+
+        logger.info("recording %s %s", hash, fid)
+        rd.hset('hash', hash, fid)
+        rd.zadd('log', {
+            fid: time.time()
+        })
+
+        logger.info("invoke indexer %s", fid)
+        indxr.index_article.delay(fid)
+        logger.info("invoke splitter %s", fid)
+        spltr.fire.delay(fid)
+
     os.unlink(fname)
+    logger.info("finish %s" % fname)
 
     return fname
